Сделал "взрыв" шарика после пересечения середины формы. Но работает покка плохо.

+++ BallGamesWinFormsApp/SalutWindowsFormsApp/SalutBall.cs
@@ -12,7 +12,7 @@ namespace SalutWindowsFormsApp
     public class SalutBall : MoveBallRandom
     {
         private float g = 0.2f;
-        public SalutBall(Form form, int centerX, int centerY) : base(form)
+        public SalutBall(Form form, float centerX, float centerY) : base(form)
         {
             radius = 5;
             brush = Brushes.Green;
